#Se agrega listado de carrito

+++ app/src/main/java/com/example/proydami/Models/Producto.kt
@@ -3,7 +3,7 @@ package com.example.proydami.Models
 import java.io.Serializable
 
 class Producto(
-    var id: Int? = 0,
+    var id: Long? = 0,
     var nombre: String? = null,
     var desc: String? = null,
     var precio: Double? = 0.0,
